requirements.txt
requests==2.25.1
beautifulsoup4==4.9.3
python_magic==0.4.24
